requirements.txt
fastapi==0.103.2
pandas==2.1.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
pydantic==2.7.3
#pip install pydantic[email]
